refactor(influxql): move the influxql package to internal
This is another package that could have outside application but is
mostly used in low usage packages. In order to restrict its usage and
restrict our own API in case of needed changes, I'm moving this to an
internal package.
